added guides abm, comments, images, tags, etc

+++ lorecodex/src/main/java/com/lorecodex/backend/controller/AdminController.java
@@ -31,7 +31,7 @@ public class AdminController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
         userService.deleteUser(id);
         return ResponseEntity.noContent().build();
     }
